+++ anthem/cli.py
@@ -33,6 +33,11 @@ def main():
         'to ctx (the the Anthem context), much like running python -i'
     )
     parser.add_argument(
+        '-q', '--quiet',
+        action='store_true',
+        help='Quiet the logs, which is sad for the songs.',
+    )
+    parser.add_argument(
         'odoo-args',
         nargs=argparse.REMAINDER,
         help='command line arguments to be passed on to Odoo, like -c for a '
@@ -40,9 +45,10 @@ def main():
         'the end of the arguments'
     )
     args = parser.parse_args()
-    odoo_args = vars(args)['odoo-args']
     print(args)
-    run(odoo_args, args.target, args.interactive)
+    odoo_args = vars(args)['odoo-args']
+    options = Options(interactive=args.interactive, quiet=args.quiet)
+    run(odoo_args, args.target, options)
 
 
 def banner():
@@ -58,13 +64,20 @@ def banner():
     return b
 
 
-def run(odoo_args, target, interactive):
-    with Context(odoo_args) as ctx:
+class Options(object):
+
+    def __init__(self, interactive=False, quiet=False):
+        self.interactive = interactive
+        self.quiet = quiet
+
+
+def run(odoo_args, target, options):
+    with Context(odoo_args, options) as ctx:
         mod_name, func_name = target.split('::')
         module = importlib.import_module(mod_name)
         func = getattr(module, func_name)
         func(ctx)
-        if interactive:
+        if options.interactive:
             console = code.InteractiveConsole(locals={'ctx': ctx})
             import rlcompleter  # noqa
             import readline
@@ -74,8 +87,9 @@ def run(odoo_args, target, interactive):
 
 
 class Context(object):
-    def __init__(self, odoo_args):
+    def __init__(self, odoo_args, options):
         self.env = self._build_odoo_env(odoo_args)
+        self.options = options
         self._log = LogIndent()
 
     def __enter__(self):
@@ -108,5 +122,8 @@ class Context(object):
 
     @contextmanager
     def log(self, name, timing=True):
-        with self._log.display(name, timing=timing):
+        if self.options.quiet:
             yield
+        else:
+            with self._log.display(name, timing=timing):
+                yield