+++ HISTORY.rst
@@ -6,6 +6,10 @@ Release History
 Unreleased
 ++++++++++
 
+**Fixes**
+
+- Set openerp's loglevel to ERROR, its logs clutter anthem's own outputs
+
 0.2.0 (2016-07-22)
 ++++++++++++++++++