+++ HISTORY.rst
@@ -15,6 +15,10 @@ Unreleased
 * Allow to use Ctrl-c to stop anthem.
 
 
+**Features**
+
+- Add --quiet mode
+
 0.2.0 (2016-07-22)
 ++++++++++++++++++